back end/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
pytesseract
opencv-python
PyPDF2
requests
